修改密码

+++ catm/src/main/java/com/lt/catm/repositories/UserRepository.java
@@ -9,6 +9,6 @@ import reactor.core.publisher.Mono;
 
 
 @Repository
-public interface UserRepository extends R2dbcRepository<User, Long> {
+public interface UserRepository extends R2dbcRepository<User, Integer> {
 
 }
